+++ public/app/features/canvas/elements/icon.tsx
@@ -2,13 +2,9 @@ import { css } from '@emotion/css';
 import { isString } from 'lodash';
 import React, { CSSProperties } from 'react';
 
+import { ColorDimensionConfig, ResourceDimensionConfig, ResourceDimensionMode } from '@grafana/schema';
 import { SanitizedSVG } from 'app/core/components/SVG/SanitizedSVG';
-import {
-  ColorDimensionConfig,
-  ResourceDimensionConfig,
-  ResourceDimensionMode,
-  getPublicOrAbsoluteUrl,
-} from 'app/features/dimensions';
+import { getPublicOrAbsoluteUrl } from 'app/features/dimensions';
 import { DimensionContext } from 'app/features/dimensions/context';
 import { ColorDimensionEditor, ResourceDimensionEditor } from 'app/features/dimensions/editors';
 import { APIEditorConfig, callApi } from 'app/plugins/panel/canvas/editor/APIEditor';
