+++ public/app/features/dimensions/color.ts
@@ -6,8 +6,9 @@ import {
   GrafanaTheme2,
   getFieldConfigWithMinMax,
 } from '@grafana/data';
+import { ColorDimensionConfig } from '@grafana/schema';
 
-import { ColorDimensionConfig, DimensionSupplier } from './types';
+import { DimensionSupplier } from './types';
 import { findField, getLastNotNullFieldValue } from './utils';
 
 //---------------------------------------------------------
