+++ public/app/features/dimensions/context.ts
@@ -1,13 +1,13 @@
 import { PanelData } from '@grafana/data/src';
-
 import {
   ColorDimensionConfig,
-  DimensionSupplier,
   ResourceDimensionConfig,
   ScalarDimensionConfig,
   ScaleDimensionConfig,
   TextDimensionConfig,
-} from './types';
+} from '@grafana/schema';
+
+import { DimensionSupplier } from './types';
 
 export interface DimensionContext {
   getColor(color: ColorDimensionConfig): DimensionSupplier<string>;
