+++ public/app/features/dimensions/editors/ScalarDimensionEditor.tsx
@@ -2,11 +2,12 @@ import { css } from '@emotion/css';
 import React, { useCallback } from 'react';
 
 import { FieldType, GrafanaTheme2, SelectableValue, StandardEditorProps } from '@grafana/data';
+import { ScalarDimensionMode, ScalarDimensionConfig } from '@grafana/schema';
 import { InlineField, InlineFieldRow, RadioButtonGroup, Select, useStyles2 } from '@grafana/ui';
 import { useFieldDisplayNames, useSelectOptions } from '@grafana/ui/src/components/MatchersUI/utils';
 import { NumberInput } from 'app/core/components/OptionsUI/NumberInput';
 
-import { ScalarDimensionConfig, ScalarDimensionMode, ScalarDimensionOptions } from '../types';
+import { ScalarDimensionOptions } from '../types';
 
 const fixedValueOption: SelectableValue<string> = {
   label: 'Fixed value',
