+++ public/app/features/dimensions/editors/ScaleDimensionEditor.tsx
@@ -2,12 +2,13 @@ import { css } from '@emotion/css';
 import React, { useCallback, useMemo } from 'react';
 
 import { GrafanaTheme2, SelectableValue, StandardEditorProps } from '@grafana/data';
+import { ScaleDimensionConfig } from '@grafana/schema';
 import { InlineField, InlineFieldRow, Select, useStyles2 } from '@grafana/ui';
 import { useFieldDisplayNames, useSelectOptions } from '@grafana/ui/src/components/MatchersUI/utils';
 import { NumberInput } from 'app/core/components/OptionsUI/NumberInput';
 
 import { validateScaleOptions, validateScaleConfig } from '../scale';
-import { ScaleDimensionConfig, ScaleDimensionOptions } from '../types';
+import { ScaleDimensionOptions } from '../types';
 
 const fixedValueOption: SelectableValue<string> = {
   label: 'Fixed value',
