+++ public/app/features/dimensions/resource.test.ts
@@ -1,5 +1,6 @@
+import { ResourceDimensionMode } from '@grafana/schema';
+
 import { getResourceDimension } from './resource';
-import { ResourceDimensionMode } from './types';
 
 describe('getResourceDimension', () => {
   const publicPath = '/public/';
