+++ public/app/features/dimensions/resource.ts
@@ -1,6 +1,7 @@
 import { DataFrame } from '@grafana/data';
+import { ResourceDimensionConfig, ResourceDimensionMode } from '@grafana/schema';
 
-import { DimensionSupplier, ResourceDimensionConfig, ResourceDimensionMode } from './types';
+import { DimensionSupplier } from './types';
 import { findField, getLastNotNullFieldValue } from './utils';
 
 //---------------------------------------------------------
