+++ public/app/features/dimensions/scalar.ts
@@ -1,6 +1,7 @@
 import { DataFrame, Field } from '@grafana/data';
+import { ScalarDimensionConfig, ScalarDimensionMode } from '@grafana/schema';
 
-import { DimensionSupplier, ScalarDimensionConfig, ScalarDimensionMode } from './types';
+import { DimensionSupplier } from './types';
 import { findField, getLastNotNullFieldValue } from './utils';
 
 //---------------------------------------------------------
