+++ public/app/features/dimensions/scale.ts
@@ -1,10 +1,9 @@
 import { DataFrame, Field } from '@grafana/data';
 import { getMinMaxAndDelta } from '@grafana/data/src/field/scale';
+import { ScaleDimensionConfig, ScaleDimensionMode } from '@grafana/schema';
 
-import { ScaleDimensionConfig, DimensionSupplier, ScaleDimensionOptions } from './types';
+import { DimensionSupplier, ScaleDimensionOptions } from './types';
 import { findField, getLastNotNullFieldValue } from './utils';
-
-import { ScaleDimensionMode } from '.';
 
 //---------------------------------------------------------
 // Scale dimension
@@ -43,7 +42,7 @@ export function getScaledDimensionForField(
   }
 
   let scaled = (percent: number) => config.min + percent * delta;
-  if (mode === ScaleDimensionMode.Quadratic) {
+  if (mode === ScaleDimensionMode.Quad) {
     const maxArea = Math.PI * (config.max / 2) ** 2;
     const minArea = Math.PI * (config.min / 2) ** 2;
     const deltaArea = maxArea - minArea;
