+++ public/app/features/dimensions/text.ts
@@ -1,6 +1,7 @@
 import { DataFrame, Field, FieldType, formattedValueToString } from '@grafana/data';
+import { TextDimensionConfig, TextDimensionMode } from '@grafana/schema';
 
-import { DimensionSupplier, TextDimensionConfig, TextDimensionMode } from './types';
+import { DimensionSupplier } from './types';
 import { findField, getLastNotNullFieldValue } from './utils';
 
 //---------------------------------------------------------
