+++ public/app/features/dimensions/types.ts
@@ -1,9 +1,5 @@
 import { Field } from '@grafana/data';
-
-export interface BaseDimensionConfig<T = any> {
-  fixed: T;
-  field?: string;
-}
+import { TextDimensionConfig, TextDimensionMode } from '@grafana/schema';
 
 export interface DimensionSupplier<T = any> {
   /**
@@ -32,33 +28,12 @@ export interface DimensionSupplier<T = any> {
   get: (index: number) => T;
 }
 
-export enum ScaleDimensionMode {
-  Linear = 'linear',
-  Quadratic = 'quad',
-}
-
-/** This will map the field value% to a scaled value within the range */
-export interface ScaleDimensionConfig extends BaseDimensionConfig<number> {
-  min: number;
-  max: number;
-}
-
 /** Places that use the value */
 export interface ScaleDimensionOptions {
   min: number;
   max: number;
   step?: number;
   hideRange?: boolean; // false
-}
-
-export enum ScalarDimensionMode {
-  Mod = 'mod',
-  Clamped = 'clamped',
-}
-export interface ScalarDimensionConfig extends BaseDimensionConfig<number> {
-  mode: ScalarDimensionMode;
-  min: number;
-  max: number;
 }
 
 export interface ScalarDimensionOptions {
@@ -70,24 +45,11 @@ export interface TextDimensionOptions {
   // anything?
 }
 
-export enum TextDimensionMode {
-  Fixed = 'fixed',
-  Field = 'field',
-  Template = 'template',
-}
-
-export interface TextDimensionConfig extends BaseDimensionConfig<string> {
-  mode: TextDimensionMode;
-}
-
 export const defaultTextConfig: TextDimensionConfig = Object.freeze({
   fixed: '',
   mode: TextDimensionMode.Field,
   field: '',
 });
-
-/** Use the color value from field configs */
-export interface ColorDimensionConfig extends BaseDimensionConfig<string> {}
 
 /** Places that use the value */
 export interface ResourceDimensionOptions {
@@ -97,18 +59,6 @@ export interface ResourceDimensionOptions {
   placeholderValue?: string;
   // If you want your icon to be driven by value of a field
   showSourceRadio?: boolean;
-}
-
-export enum ResourceDimensionMode {
-  Fixed = 'fixed',
-  Field = 'field',
-  Mapping = 'mapping',
-  // pattern? uses field in the pattern
-}
-
-/** Get the path to a resource (URL) */
-export interface ResourceDimensionConfig extends BaseDimensionConfig<string> {
-  mode: ResourceDimensionMode;
 }
 
 export enum ResourceFolderName {
