+++ public/app/plugins/panel/geomap/style/types.ts
@@ -2,14 +2,14 @@ import { Style } from 'ol/style';
 
 import {
   ColorDimensionConfig,
-  DimensionSupplier,
   ResourceDimensionConfig,
   ResourceDimensionMode,
   ScaleDimensionConfig,
   ScalarDimensionConfig,
   ScalarDimensionMode,
   TextDimensionConfig,
-} from 'app/features/dimensions';
+} from '@grafana/schema';
+import { DimensionSupplier } from 'app/features/dimensions';
 
 export enum GeometryTypeId {
   Point = 'point',
