+++ public/app/plugins/panel/geomap/style/utils.ts
@@ -1,5 +1,5 @@
 import { config } from '@grafana/runtime';
-import { TextDimensionMode } from 'app/features/dimensions';
+import { TextDimensionMode } from '@grafana/schema';
 
 import { getMarkerMaker } from './markers';
 import { defaultStyleConfig, StyleConfig, StyleConfigFields, StyleConfigState } from './types';
