+++ public/app/plugins/panel/icon/IconPanel.tsx
@@ -1,21 +1,23 @@
 import React, { Component } from 'react';
 
 import { PanelProps } from '@grafana/data';
+import {
+  ColorDimensionConfig,
+  ResourceDimensionConfig,
+  ScalarDimensionConfig,
+  ScaleDimensionConfig,
+  TextDimensionConfig,
+} from '@grafana/schema';
 import { HorizontalConstraint, VerticalConstraint } from 'app/features/canvas';
 import { iconItem } from 'app/features/canvas/elements/icon';
 import { ElementState } from 'app/features/canvas/runtime/element';
 import {
-  ColorDimensionConfig,
   DimensionContext,
   getColorDimensionFromData,
   getResourceDimensionFromData,
   getScalarDimensionFromData,
   getScaleDimensionFromData,
   getTextDimensionFromData,
-  ResourceDimensionConfig,
-  ScalarDimensionConfig,
-  ScaleDimensionConfig,
-  TextDimensionConfig,
 } from 'app/features/dimensions';
 
 import { PanelOptions } from './models.gen';
