+++ public/app/plugins/panel/xychart/AutoEditor.tsx
@@ -11,7 +11,7 @@ import {
 import { Field, IconButton, Select, useStyles2 } from '@grafana/ui';
 
 import { getXYDimensions, isGraphable } from './dims';
-import { XYDimensionConfig, XYChartOptions } from './models.gen';
+import { XYDimensionConfig, PanelOptions } from './panelcfg.gen';
 
 interface XYInfo {
   numberFields: Array<SelectableValue<string>>;
@@ -19,11 +19,7 @@ interface XYInfo {
   yFields: Array<SelectableValue<boolean>>;
 }
 
-export const AutoEditor = ({
-  value,
-  onChange,
-  context,
-}: StandardEditorProps<XYDimensionConfig, any, XYChartOptions>) => {
+export const AutoEditor = ({ value, onChange, context }: StandardEditorProps<XYDimensionConfig, any, PanelOptions>) => {
   const frameNames = useMemo(() => {
     if (context?.data?.length) {
       return context.data.map((f, idx) => ({
