+++ public/app/plugins/panel/xychart/ManualEditor.tsx
@@ -7,7 +7,7 @@ import { FieldNamePicker } from '@grafana/ui/src/components/MatchersUI/FieldName
 import { LayerName } from 'app/core/components/Layers/LayerName';
 import { ColorDimensionEditor, ScaleDimensionEditor } from 'app/features/dimensions/editors';
 
-import { PanelOptions, ScatterSeriesConfig, defaultScatterFieldConfig } from './types';
+import { PanelOptions, ScatterSeriesConfig, defaultScatterFieldConfig } from './panelcfg.gen';
 
 export const ManualEditor = ({
   value,
