+++ public/app/plugins/panel/xychart/TooltipView.tsx
@@ -14,7 +14,7 @@ import { LinkButton, usePanelContext, useStyles2, VerticalGroup, VizTooltipOptio
 import { findField } from 'app/features/dimensions';
 import { getFieldLinksForExplore } from 'app/features/explore/utils/links';
 
-import { ScatterSeriesConfig, SeriesMapping } from './models.gen';
+import { ScatterSeriesConfig, SeriesMapping } from './panelcfg.gen';
 import { ScatterSeries } from './types';
 
 interface YValue {
