+++ public/app/plugins/panel/xychart/module.tsx
@@ -5,7 +5,7 @@ import { AutoEditor } from './AutoEditor';
 import { ManualEditor } from './ManualEditor';
 import { XYChartPanel2 } from './XYChartPanel2';
 import { getScatterFieldConfig } from './config';
-import { defaultScatterFieldConfig, PanelOptions, ScatterFieldConfig } from './types';
+import { PanelOptions, ScatterFieldConfig, defaultScatterFieldConfig } from './panelcfg.gen';
 
 export const plugin = new PanelPlugin<PanelOptions, ScatterFieldConfig>(XYChartPanel2)
   .useFieldConfig(getScatterFieldConfig(defaultScatterFieldConfig))
