+++ public/app/plugins/panel/xychart/panelcfg.gen.ts
@@ -34,16 +34,16 @@ export const defaultXYDimensionConfig: Partial<XYDimensionConfig> = {
 };
 
 export interface ScatterFieldConfig extends common.HideableFieldConfig, common.AxisConfig {
+  fillOpacity?: number; // TODO: 0-1, default to 0.5
   label?: common.VisibilityMode;
   labelValue?: common.TextDimensionConfig;
   lineColor?: common.ColorDimensionConfig;
   lineStyle?: common.LineStyle;
   lineWidth?: number;
-  opacity?: number; // TODO: 0-1, default to 0.5
   pointColor?: common.ColorDimensionConfig;
   pointSize?: common.ScaleDimensionConfig;
+  pointSymbol?: common.ResourceDimensionConfig;
   show?: ScatterShow;
-  symbol?: common.ResourceDimensionConfig;
 }
 
 export const defaultScatterFieldConfig: Partial<ScatterFieldConfig> = {
