+++ public/app/plugins/panel/xychart/scatter.ts
@@ -13,28 +13,23 @@ import {
 } from '@grafana/data';
 import { alpha } from '@grafana/data/src/themes/colorManipulator';
 import { config } from '@grafana/runtime';
-import { AxisPlacement, ScaleDirection, ScaleOrientation, VisibilityMode } from '@grafana/schema';
-import { UPlotConfigBuilder } from '@grafana/ui';
-import { FacetedData, FacetSeries } from '@grafana/ui/src/components/uPlot/types';
 import {
-  findFieldIndex,
-  getScaledDimensionForField,
+  AxisPlacement,
+  ScaleDirection,
+  ScaleOrientation,
+  VisibilityMode,
   ScaleDimensionConfig,
   ScaleDimensionMode,
-} from 'app/features/dimensions';
+} from '@grafana/schema';
+import { UPlotConfigBuilder } from '@grafana/ui';
+import { FacetedData, FacetSeries } from '@grafana/ui/src/components/uPlot/types';
+import { findFieldIndex, getScaledDimensionForField } from 'app/features/dimensions';
 
 import { pointWithin, Quadtree, Rect } from '../barchart/quadtree';
 
 import { isGraphable } from './dims';
-import {
-  DimensionValues,
-  ScatterFieldConfig,
-  defaultScatterFieldConfig,
-  ScatterHoverCallback,
-  ScatterSeries,
-  PanelOptions,
-  ScatterShow,
-} from './types';
+import { ScatterFieldConfig, defaultScatterFieldConfig, PanelOptions, ScatterShow } from './panelcfg.gen';
+import { DimensionValues, ScatterHoverCallback, ScatterSeries } from './types';
 
 export interface ScatterPanelInfo {
   error?: string;
@@ -142,7 +137,7 @@ function getScatterSeries(
       const s = getScaledDimensionForField(
         frame.fields[dims.pointSizeIndex!],
         dims.pointSizeConfig!,
-        ScaleDimensionMode.Quadratic
+        ScaleDimensionMode.Quad
       );
       const vals = Array(frame.length);
       for (let i = 0; i < frame.length; i++) {
@@ -295,7 +290,7 @@ interface DrawBubblesOpts {
   };
 }
 
-//const prepConfig: UPlotConfigPrepFnXY<XYChartOptions> = ({ frames, series, theme }) => {
+//const prepConfig: UPlotConfigPrepFnXY<PanelOptions> = ({ frames, series, theme }) => {
 const prepConfig = (
   getData: () => DataFrame[],
   scatterSeries: ScatterSeries[],
