+++ public/app/plugins/panel/xychart/types.ts
@@ -1,14 +1,6 @@
 import { DataFrame, Field, FieldColorMode } from '@grafana/data';
 import { LineStyle, ScaleDimensionConfig, VisibilityMode } from '@grafana/schema';
 import { VizLegendItem } from '@grafana/ui';
-import { DimensionSupplier } from 'app/features/dimensions';
-
-import {
-  ScatterFieldConfig as GeneratedScatterFieldConfig,
-  ScatterSeriesConfig as GeneratedScatterSeriesConfig,
-  defaultScatterFieldConfig as generatedDefaultScatterFieldConfig,
-  PanelOptions as GeneratedPanelOptions,
-} from './panelcfg.gen';
 
 /**
  * @internal
@@ -23,20 +15,6 @@ export interface ScatterHoverEvent {
 }
 
 export type ScatterHoverCallback = (evt?: ScatterHoverEvent) => void;
-
-export interface ScatterFieldConfig extends GeneratedScatterFieldConfig {
-  pointSymbol?: DimensionSupplier<string>;
-}
-
-export const defaultScatterFieldConfig: Partial<ScatterFieldConfig> = {
-  ...generatedDefaultScatterFieldConfig,
-};
-
-export interface ScatterSeriesConfig extends ScatterFieldConfig, GeneratedScatterSeriesConfig {}
-
-export interface PanelOptions extends Omit<GeneratedPanelOptions, 'series'> {
-  series: ScatterSeriesConfig[];
-}
 
 // Using field where we will need formatting/scale/axis info
 // Use raw or DimensionValues when the values can be used directly
@@ -72,5 +50,3 @@ export interface ScatterSeries {
     };
   };
 }
-
-export { ScatterShow, SeriesMapping, XYDimensionConfig } from './panelcfg.gen';
